modified project

+++ api/src/main/java/ru/gb/shop/api/CartDto.java
@@ -1,17 +1,18 @@
 package ru.gb.shop.api;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 
 public class CartDto {
     private List<CartItemDto> items;
-    private double totalPrice;
+    private BigDecimal totalPrice;
 
     public List<CartItemDto> getItems() {
         return items;
     }
 
-    public double getTotalPrice() {
+    public BigDecimal getTotalPrice() {
         return totalPrice;
     }
 
@@ -19,11 +20,11 @@ public class CartDto {
         this.items = items;
     }
 
-    public void setTotalPrice(double totalPrice) {
+    public void setTotalPrice(BigDecimal totalPrice) {
         this.totalPrice = totalPrice;
     }
 
-    public CartDto(List<CartItemDto> items, double totalPrice) {
+    public CartDto(List<CartItemDto> items, BigDecimal totalPrice) {
         this.items = items;
         this.totalPrice = totalPrice;
     }
