+++ api/src/main/java/ru/gb/shop/api/ProductDto.java
@@ -1,9 +1,11 @@
 package ru.gb.shop.api;
 
+import java.math.BigDecimal;
+
 public class ProductDto {
     private Long id;
     private String title;
-    private double price;
+    private BigDecimal price;
     private String categoryTitle;
 
     public Long getId() {
@@ -22,11 +24,11 @@ public class ProductDto {
         this.title = title;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
 
@@ -38,7 +40,7 @@ public class ProductDto {
         this.categoryTitle = categoryTitle;
     }
 
-    public ProductDto(Long id, String title, double price, String categoryTitle) {
+    public ProductDto(Long id, String title, BigDecimal price, String categoryTitle) {
         this.id = id;
         this.title = title;
         this.price = price;
