+++ core-service/src/main/java/ru/gb/shop/core/entities/OrderItem.java
@@ -6,6 +6,7 @@ import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -29,10 +30,10 @@ public class OrderItem {
     private int quantity;
 
     @Column(name = "price_per_product")
-    private double pricePerProduct;
+    private BigDecimal pricePerProduct;
 
     @Column(name = "price")
-    private double price;
+    private BigDecimal price;
 
     @CreationTimestamp
     @Column(name = "created_at")
@@ -41,7 +42,7 @@ public class OrderItem {
     @Column(name = "updated_at")
     private LocalDateTime updatedAt;
 
-    public OrderItem(Product product, Order order, int quantity, double pricePerProduct, double price) {
+    public OrderItem(Product product, Order order, int quantity, BigDecimal pricePerProduct, BigDecimal price) {
         this.product = product;
         this.order = order;
         this.quantity = quantity;
