+++ core-service/src/main/java/ru/gb/shop/core/specifications/ProductSpecifications.java
@@ -4,10 +4,10 @@ import org.springframework.data.jpa.domain.Specification;
 import ru.gb.shop.core.entities.Product;
 
 public class ProductSpecifications {
-    public static Specification<Product> costGreaterOrEqualsThan(Double minCost) {
+    public static Specification<Product> costGreaterOrEqualsThan(Integer minCost) {
         return (root, criteriaQuery, criteriaBuilder) -> criteriaBuilder.greaterThanOrEqualTo(root.get("cost"), minCost);
     }
-    public static Specification<Product> costLessOrEqualsThan(Double maxCost) {
+    public static Specification<Product> costLessOrEqualsThan(Integer maxCost) {
         return (root, criteriaQuery, criteriaBuilder) -> criteriaBuilder.lessThanOrEqualTo(root.get("cost"), maxCost);
     }
     public static Specification<Product> titleLike(String titlePart) {
